+++ src/bot/tests/parse_test.ts
@@ -1,15 +1,23 @@
 import { assertEquals, assertThrows } from "@std/assert";
-import { blockLocationParamType } from "../lib/command.ts";
+import {
+  blockLocationParamType,
+  OptionalCommandParameter,
+} from "../lib/command.ts";
 import { lexCommandInput, parseCommand } from "../lib/parser.ts";
 import {
   booleanParamType,
+  type CommandParameter,
   floatParamType,
+  integerParamType,
   jsonParamType,
   type Location,
   locationParamType,
+  MissingArgumentError,
+  PartialArgumentError,
   stringParamType,
-  TypeError,
   SyntaxError,
+  TooManyArgumentsError,
+  TypeError,
 } from "@bedrock-ws/bot";
 import { assertInstanceOf } from "@std/assert/instance-of";
 
@@ -57,7 +65,75 @@ Deno.test("lexer", async (t) => {
   await t.step("empty input", () => {
     const error = assertThrows(() => lexCommandInput(""));
     assertInstanceOf(error, SyntaxError);
-  })
+  });
+});
+
+Deno.test("default value", () => {
+  const mandatoryParameters: CommandParameter<never>[] = [];
+  const optionalParameters: OptionalCommandParameter<number>[] = [
+    { name: "value", type: integerParamType, default: { value: 1 } },
+    {
+      name: "callback",
+      type: integerParamType,
+      default: { factory: () => 2 },
+    },
+    { name: "raw", type: integerParamType, default: { raw: ["3"] } },
+    { name: "none", type: integerParamType },
+  ];
+  const args: string[] = [];
+  const values = parseCommand(mandatoryParameters, optionalParameters, args);
+  assertEquals(values, [1, 2, 3, undefined]);
+});
+
+Deno.test("argument amount mismatch", async (t) => {
+  const mandatoryParameters = [
+    { name: "a", type: stringParamType },
+    { name: "b", type: stringParamType },
+  ];
+  const optionalParameters: CommandParameter<never>[] = [];
+
+  await t.step("too many", () => {
+    const args = ["A", "B", "C"];
+    const error = assertThrows(() =>
+      parseCommand(mandatoryParameters, optionalParameters, args)
+    );
+    assertInstanceOf(error, TooManyArgumentsError);
+  });
+
+  await t.step("too few", () => {
+    const args = ["A"];
+    const error = assertThrows(() =>
+      parseCommand(mandatoryParameters, optionalParameters, args)
+    );
+    assertInstanceOf(error, MissingArgumentError);
+  });
+
+  await t.step("only partially provided", async (t) => {
+    await t.step("mandatory parameter", () => {
+      const mandatoryParameters: CommandParameter<never>[] = [];
+      const optionalParameters = [{
+        name: "a",
+        type: blockLocationParamType,
+      }];
+      const args = ["1", "2"];
+      const error = assertThrows(() =>
+        parseCommand(mandatoryParameters, optionalParameters, args)
+      );
+      assertInstanceOf(error, PartialArgumentError);
+    });
+    await t.step("optional parameter", () => {
+      const mandatoryParameters = [{
+        name: "a",
+        type: blockLocationParamType,
+      }];
+      const optionalParameters: CommandParameter<never>[] = [];
+      const args = ["1", "2"];
+      const error = assertThrows(() =>
+        parseCommand(mandatoryParameters, optionalParameters, args)
+      );
+      assertInstanceOf(error, PartialArgumentError);
+    });
+  });
 });
 
 Deno.test("string", () => {
@@ -208,6 +284,17 @@ Deno.test("float", () => {
   const input = [value.toString()];
   const args = parseCommand(
     [{ name: "foo", type: floatParamType }],
+    [],
+    input,
+  );
+  assertEquals(args, [value]);
+});
+
+Deno.test("integer", () => {
+  const value = 42;
+  const input = [value.toString()];
+  const args = parseCommand(
+    [{ name: "foo", type: integerParamType }],
     [],
     input,
   );